requirements.txt
google-genai
qdrant-client
python-dotenv
streamlit
pypdf
